Move schedule the the p3 module and a new url.
