+++ p3/urls.py
@@ -33,8 +33,6 @@ urlpatterns = [
 
 
 urlpatterns += [
-    url(r'^schedule/$', p3_views.schedule, name='p3-schedule'),
-    url(r'^schedule/(?P<day>\d+)-(?P<month>\w+)$', p3_views.schedule, name='p3-schedule'),
     # url(r'^schedule/(?P<conference>[\w-]+).ics$', p3_views.schedule_ics, name='p3-schedule-ics'),
 
     # url(r'^schedule/(?P<conference>[\w-]+)/my-schedule/$',
