+++ pycon/urls.py
@@ -28,6 +28,7 @@ from conference.talk_voting import urlpatterns as talk_voting_urls
 from conference.user_panel import urlpatterns as user_panel_urls
 from conference.talks import urlpatterns as talks_urls
 from conference.profiles import urlpatterns as profiles_urls
+from conference.schedule import urlpatterns as schedule_urls
 
 admin.autodiscover()
 admin.site.index_template = 'p3/admin/index.html'
@@ -44,6 +45,7 @@ urlpatterns = [
     url(r'^talks/', include(talks_urls, namespace="talks")),
     url(r'^profiles/', include(profiles_urls, namespace="profiles")),
     url(r'^talk-voting/', include(talk_voting_urls, namespace="talk_voting")),
+    url(r'^schedule/', include(schedule_urls, namespace="schedule")),
     url(r'^news/', news_list, name="news"),
     url(r'^accounts/', include('assopy.urls')),
     url(r'^admin/filebrowser/', include(fsite.urls)),
