+++ templates/conference/talk.html
@@ -85,7 +85,7 @@
                         {% comment %}
                         (expected attendance {{ e.expected_attendance.expected|floatformat:"0" }} people)
                         {% endcomment %}
-                        <a href="{% url "p3-schedule" conference=e.schedule.conference %}"><b>See schedule</b></a>
+                        <a href="{% url "schedule:schedule" %}"><b>See schedule</b></a>
                     </section>
                     {% endif %}
                 {% endfor %}
